# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "\u003cadd description here\u003e",
        "title": "ORCHID_FOOTWARES",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:9080",
    "basePath": "/",
    "paths": {
        "/admin/addcoupen": {
            "post": {
                "description": "Admin can add coupen here",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin",
                    "coupen management"
                ],
                "summary": "ADD COUPEN",
                "operationId": "add-coupen",
                "parameters": [
                    {
                        "type": "string",
                        "description": "code",
                        "name": "code",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "minimumamount",
                        "name": "minimumamount",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "amount",
                        "name": "amount",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
...
        "/admin/listcoupen": {
            "get": {
                "description": "Admin can list coupen here",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin",
                    "coupen management"
                ],
                "summary": "LIST COUPEN",
                "operationId": "list-coupen",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
...
        "/user/aboutme": {
            "get": {
                "description": "user can view profile",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User",
                    "profile management"
                ],
                "summary": "VIEW  PROFILE",
                "operationId": "view-profile",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
...
        "/user/addorder": {
            "post": {
                "description": "user can add order here",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User",
                    "order management"
                ],
                "summary": "ADD ORDER",
                "operationId": "add-order",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
...
        "/user/applycoupens": {
            "post": {
                "description": "user can apply coupen here",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User",
                    "coupen"
                ],
                "summary": "APPLY COUPEN",
                "operationId": "applycoupen",
                "parameters": [
                    {
                        "type": "string",
                        "description": "coupenname",
                        "name": "coupenname",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/user/cancelorder": {
            "delete": {
                "description": "user can cancel order here",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User",
                    "order management"
                ],
                "summary": "CANCEL ORDER",
                "operationId": "cancel-order",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/user/changepassword": {
            "post": {
                "description": "User can change password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User",
                    "Change password"
                ],
                "summary": "Change password",
                "operationId": "change-password",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
...
        "/user/deleteprofile": {
            "delete": {
                "description": "user can delete profile",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User",
                    "profile management"
                ],
                "summary": "DELETE  PROFILE",
                "operationId": "delete-profile",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
...
        "/user/editmyprofile": {
            "patch": {
                "description": "Edit user profile with the provided data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User",
                    "profile management"
                ],
                "summary": "Edit user profile",
                "operationId": "edit-profile",
                "parameters": [
                    {
                        "description": "informations",
                        "name": "EditprofileResponse",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.EditprofileResponse"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/user/listcoupens": {
            "get": {
                "description": "user can list coupen here",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User",
                    "coupen"
                ],
                "summary": "LIST COUPEN",
                "operationId": "listcoupen",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
...
        "/user/showorder": {
            "get": {
                "description": "user can view order here",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User",
                    "order management"
                ],
                "summary": "SHOW ORDER",
                "operationId": "show-order",
                "parameters": [
                    {
                        "type": "string",
                        "description": "page",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "perpage",
                        "name": "perpage",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.EditprofileResponse": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "phone_Number": {
                    "type": "string"
                },
                "user_Name": {
                    "type": "string"
                }
            }
        },
...
    }
}